Update README.md
Added 2013 MBPr to the supported models. It works on my 13" model and I saw in the issues that the 15" model has been supported for a while, too.


In one of those issues I saw that you where working on a thumb-detection algorithm. Shall this be added as a separate point in the Roadmap, or as more information under Signal Processing Algorithm?

And, last but not least: thank you! While I was mostly looking for thumb detection, this already improved the touchpad behaviour quite a bit!

Your 'buy me a coffee' link is dead, by the way.