T2 Devices: initial km touchpad project.
Although this branch is named as "T2", it may be used for
moving current USB um project into km in the future. Currently,
I am only testing basic functions on a Magic Trackpad 2, as the
T2 trackpad struct is not fully known (I guess pretty identical).
It is capable to initialize the USB device and bring it into the
desired mode for now.

T2 VHCI device does not play well with UMDF USB driver framework,
the NativeUSB dispatcher fails when retrieving USB configuration.