USB: Extending input report identifier to uint32_t
Part of #227

We actually report the finger origin to Windows on Magic Trackpad 2 and it works good. We didn't do that on MacBook but this will eventually happen in the future. Since the origin identifier reported by MacBook trackpad is not very stable, we will track fingers on our own. To make the process easier, the report identifier field has been extended to uint32_t.